refactor: move profile update components to correct feature folder
