+++ src/features/auth/components/RegisterForm.tsx
@@ -6,7 +6,6 @@ import z from "zod";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
-import toast from "react-hot-toast";
 import { cn } from "@/shared/lib/utils";
 import {
   Form,
